requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
lxml==4.5.0
PySocks==1.7.1
pyTelegramBotAPI 3.6.7
requests==2.23.0
urllib3==1.25.8